bsdiff: Don't install delta_generator tools on the device.
Update Engine's delta_generator requires the bsdiff binary to be
present on the host when the au-generator.zip is created with the
generate_au_zip.py script. This .zip file will later run on the
builders. Nevertheless, the bsdiff binary is not required on the
ChromeOS devices, only bspatch is.

This patch removes the bsdiff and its dependencies from the device
saving a few KB.

BUG=chromium:372644
TEST=Manual test:
TEST=emerge-link --unmerge libdivsufsort && emerge-link bsdiff # Both work, bsdiff not present on link
TEST=sudo emerge bsdiff # Works, bsdiff present on the host.

Change-Id: Ie2823c796185d616523fe8bf79a7c4c91fa3a0c0
Reviewed-on: https://chromium-review.googlesource.com/199466
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>